Controle_Access_Titres/srv-controle/requirements.txt
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.5.0
Flask==3.0.0
Flask-WTF==1.2.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
paho-mqtt==1.6.1
pypng==0.20220715.0
requests==2.31.0
setuptools==69.0.3
typing_extensions==4.9.0
urllib3==2.1.0
Werkzeug==3.0.1
wheel==0.42.0

